Fixed #168 Adverserial Search Agent (#170)
* Fixing #168 the Failing Search Agent

Fixed Alpha-Beta pruning agent, now the Tic-Tac-Toe computer agent can never lose.

* Tree object should be local

The previous commit had the tree object in game.js declared global in the previous commit. Fixed that, as it's no longer needed for debugging.

* Fixed Permissions

Had changed the permissions in the first commit when trying to run this, changed them back.